Reused method from Apache library instead of own implementation Moved classes to different packages and sorted by purpose
