+++ src/main/java/net/masterthought/cucumber/json/support/StepResults.java
@@ -3,8 +3,6 @@ package net.masterthought.cucumber.json.support;
 import java.util.List;
 
 import net.masterthought.cucumber.json.Step;
-import net.masterthought.cucumber.util.Status;
-import net.masterthought.cucumber.util.StatusCounter;
 import net.masterthought.cucumber.util.Util;
 
 public class StepResults {
