+++ src/test/java/net/masterthought/cucumber/StepTest.java
@@ -17,7 +17,7 @@ import org.junit.Test;
 import net.masterthought.cucumber.json.Feature;
 import net.masterthought.cucumber.json.Row;
 import net.masterthought.cucumber.json.Step;
-import net.masterthought.cucumber.util.Status;
+import net.masterthought.cucumber.json.support.Status;
 
 public class StepTest {
